déplacement des fichiers en répertoires logiques
