+++ src/main.py
@@ -1,0 +1,106 @@
+import pygame
+pygame.init()
+
+win = pygame.display.set_mode((0, 0))
+pygame.display.set_caption("Project March")
+
+x1 = win.get_width()
+y1 = win.get_height()
+
+
+walkRight = [pygame.image.load('R1.png'), pygame.image.load('R2.png'), pygame.image.load('R3.png'), pygame.image.load('R4.png'), pygame.image.load('R5.png'), pygame.image.load('R6.png'), pygame.image.load('R7.png'), pygame.image.load('R8.png')]
+walkLeft = [pygame.image.load('L1.png'), pygame.image.load('L2.png'), pygame.image.load('L3.png'), pygame.image.load('L4.png'), pygame.image.load('L5.png'), pygame.image.load('L6.png'), pygame.image.load('L7.png'), pygame.image.load('L8.png')]
+bg = pygame.image.load('background_image.png')
+bg =pygame.transform.scale(bg, (1920, 1080))
+char = pygame.image.load('idle outline.gif')
+jumpima = pygame.image.load('jump outline.png')
+
+
+clock = pygame.time.Clock()
+
+
+
+
+run = True
+
+class player(object):
+    def __init__(self,x,y,width,height):
+        self.x = x
+        self.y = y
+        self.width = width
+        self.height = height
+        self.vel = 5
+        self.isJump = False
+        self.left = False
+        self.right = False
+        self.walkCount = 0
+        self.jumpCount = 10
+
+    def draw(self, win):
+        if self.walkCount + 1 >= 23:
+            self.walkCount = 0
+
+        if self.left:
+            win.blit(walkLeft[self.walkCount // 3], (self.x, self.y))
+            self.walkCount += 1
+        elif self.right:
+            win.blit(walkRight[self.walkCount // 3], (self.x, self.y))
+            self.walkCount += 1
+        elif self.isJump :
+            win.blit(jumpima, (self.x, self.y))
+        else:
+            win.blit(char, (self.x, self.y))
+
+
+def redrawGameWindow():
+    win.blit(bg, (0, 0))
+    man.draw(win)
+
+    pygame.display.update()
+
+
+man = player(100, 970, 64, 64)
+run = True
+while run:
+    clock.tick(23)
+
+    for event in pygame.event.get():
+        if event.type == pygame.QUIT:
+            run = False
+
+    keys = pygame.key.get_pressed()
+
+    if keys[pygame.K_LEFT] and man.x > man.vel:
+        man.x -= man.vel
+        man.left = True
+        man.right = False
+    elif keys[pygame.K_RIGHT] and man.x < x1 - man.width - man.vel:
+        man.x += man.vel
+        man.right = True
+        man.left = False
+    else:
+        man.right = False
+        man.left = False
+        man.walkCount = 0
+
+    if not (man.isJump):
+        if keys[pygame.K_SPACE]:
+            man.isJump = True
+            man.right = False
+            man.left = False
+            man.walkCount = 0
+    else:
+        if man.jumpCount >= -10:
+            neg = 1
+            if man.jumpCount < 0:
+                neg = -1
+            man.y -= (man.jumpCount ** 2) * 0.35 * neg
+            man.jumpCount -= 1
+        else:
+            man.isJump = False
+            man.jumpCount = 10
+
+    redrawGameWindow()
+
+pygame.quit()
+
